+++ plugins/destination/clickhouse/client/spec/spec_test.go
@@ -1,0 +1,1 @@
+package spec
