+++ plugins/destination/clickhouse/queries/tables.go
@@ -1,10 +1,10 @@
 package queries
 
 import (
+	"slices"
 	"strings"
 
-	"slices"
-
+	"github.com/cloudquery/cloudquery/plugins/destination/clickhouse/client/spec"
 	"github.com/cloudquery/cloudquery/plugins/destination/clickhouse/typeconv/ch/types"
 	"github.com/cloudquery/cloudquery/plugins/destination/clickhouse/util"
 	"github.com/cloudquery/plugin-sdk/v4/schema"
@@ -27,7 +27,7 @@ func sortKeys(table *schema.Table) []string {
 	return slices.Clip(keys)
 }
 
-func CreateTable(table *schema.Table, cluster string, engine *Engine) (string, error) {
+func CreateTable(table *schema.Table, cluster string, engine *spec.Engine) (string, error) {
 	builder := strings.Builder{}
 	builder.WriteString("CREATE TABLE ")
 	builder.WriteString(tableNamePart(table.Name, cluster))
