+++ plugins/destination/clickhouse/queries/tables_test.go
@@ -5,6 +5,7 @@ import (
 	"testing"
 
 	"github.com/apache/arrow/go/v14/arrow"
+	"github.com/cloudquery/cloudquery/plugins/destination/clickhouse/client/spec"
 	"github.com/cloudquery/plugin-sdk/v4/schema"
 	"github.com/cloudquery/plugin-sdk/v4/types"
 	"github.com/stretchr/testify/require"
@@ -26,7 +27,7 @@ func TestCreateTable(t *testing.T) {
 			schema.Column{Name: "extra_inet_col", Type: types.NewInetType()},
 			schema.Column{Name: "extra_inet_arr_col", Type: arrow.ListOf(types.NewInetType())},
 		},
-	}, "", DefaultEngine())
+	}, "", spec.DefaultEngine())
 	require.NoError(t, err)
 	ensureContents(t, query, "create_table.sql")
 }
@@ -39,7 +40,7 @@ func TestCreateTableNoOrderBy(t *testing.T) {
 			schema.Column{Name: "extra_inet_col", Type: types.NewInetType()},
 			schema.Column{Name: "extra_inet_arr_col", Type: arrow.ListOf(types.NewInetType())},
 		},
-	}, "", DefaultEngine())
+	}, "", spec.DefaultEngine())
 	require.NoError(t, err)
 	ensureContents(t, query, "create_table_empty_order_by.sql")
 }
@@ -60,7 +61,7 @@ func TestCreateTableOnCluster(t *testing.T) {
 			schema.Column{Name: "extra_inet_col", Type: types.NewInetType()},
 			schema.Column{Name: "extra_inet_arr_col", Type: arrow.ListOf(types.NewInetType())},
 		},
-	}, "my_cluster", DefaultEngine())
+	}, "my_cluster", spec.DefaultEngine())
 	require.NoError(t, err)
 	ensureContents(t, query, "create_table_cluster.sql")
 }
@@ -81,7 +82,7 @@ func TestCreateTableWithEngine(t *testing.T) {
 			schema.Column{Name: "extra_inet_col", Type: types.NewInetType()},
 			schema.Column{Name: "extra_inet_arr_col", Type: arrow.ListOf(types.NewInetType())},
 		},
-	}, "", &Engine{
+	}, "", &spec.Engine{
 		Name:       "ReplicatedMergeTree",
 		Parameters: []any{"a", "b", 1, int32(2), int64(3), float32(1.2), float64(3.4), json.Number("327"), false, true},
 	})
